refactor: EventRepository 수정
