+++ src/main/java/dev/flab/studytogether/domain/event/EventRepository.java
@@ -1,7 +1,6 @@
 package dev.flab.studytogether.domain.event;
 
-import dev.flab.studytogether.domain.event.AbstractEvent;
 
 public interface EventRepository {
-    void save(AbstractEvent event);
+    void save(DomainEvent event);
 }
