+++ src/main/java/dev/flab/studytogether/domain/event/infrastructure/EventRepositoryImpl.java
@@ -1,11 +1,10 @@
-package dev.flab.studytogether.domain.event;
+package dev.flab.studytogether.domain.event.infrastructure;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import dev.flab.studytogether.domain.event.*;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
@@ -16,11 +15,10 @@ import java.util.Map;
 @Repository
 @RequiredArgsConstructor
 public class EventRepositoryImpl implements EventRepository {
-    private final Logger log = LoggerFactory.getLogger(this.getClass());
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void save(AbstractEvent event) {
+    public void save(DomainEvent event) {
         SimpleJdbcInsert jdbcInsert = new SimpleJdbcInsert(jdbcTemplate);
         jdbcInsert.withTableName("EVENT").usingGeneratedKeyColumns("EVENT_ID");
